test: Break up router_test to avoid compiler OOM under ASAN (#14844)
