+++ test/common/router/BUILD
@@ -280,8 +280,13 @@ envoy_cc_test(
 
 envoy_cc_test(
     name = "router_test",
-    srcs = ["router_test.cc"],
+    srcs = [
+        # Split to avoid compiler OOM, especially on ASAN.
+        "router_test.cc",
+        "router_2_test.cc",
+    ],
     deps = [
+        ":router_test_base_lib",
         "//source/common/buffer:buffer_lib",
         "//source/common/http:context_lib",
         "//source/common/network:application_protocol_lib",
@@ -312,6 +317,38 @@ envoy_cc_test(
         "@envoy_api//envoy/extensions/upstreams/http/tcp/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/upstreams/tcp/generic/v3:pkg_cc_proto",
         "@envoy_api//envoy/type/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_test_library(
+    name = "router_test_base_lib",
+    srcs = ["router_test_base.cc"],
+    hdrs = ["router_test_base.h"],
+    deps = [
+        "//source/common/buffer:buffer_lib",
+        "//source/common/http:context_lib",
+        "//source/common/network:application_protocol_lib",
+        "//source/common/network:utility_lib",
+        "//source/common/router:router_lib",
+        "//source/common/stream_info:uint32_accessor_lib",
+        "//source/common/upstream:upstream_includes",
+        "//source/common/upstream:upstream_lib",
+        "//source/extensions/upstreams/http/generic:config",
+        "//source/extensions/upstreams/http/http:config",
+        "//source/extensions/upstreams/http/tcp:config",
+        "//test/common/http:common_lib",
+        "//test/mocks/http:http_mocks",
+        "//test/mocks/local_info:local_info_mocks",
+        "//test/mocks/network:network_mocks",
+        "//test/mocks/router:router_mocks",
+        "//test/mocks/runtime:runtime_mocks",
+        "//test/mocks/ssl:ssl_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
+        "//test/mocks/upstream:host_mocks",
+        "//test/test_common:environment_lib",
+        "//test/test_common:simulated_time_system_lib",
+        "//test/test_common:test_runtime_lib",
+        "//test/test_common:utility_lib",
     ],
 )
